+++ crates/cdk-cli/src/sub_commands/check_spent.rs
@@ -1,11 +1,8 @@
-use std::collections::HashMap;
-
 use anyhow::Result;
-use cdk::url::UncheckedUrl;
-use cdk::wallet::Wallet;
+use cdk::wallet::MultiMintWallet;
 
-pub async fn check_spent(wallets: HashMap<UncheckedUrl, Wallet>) -> Result<()> {
-    for wallet in wallets.values() {
+pub async fn check_spent(multi_mint_wallet: &MultiMintWallet) -> Result<()> {
+    for wallet in multi_mint_wallet.get_wallets().await {
         let amount = wallet.check_all_pending_proofs().await?;
 
         println!("Amount marked as spent: {}", amount);