+++ crates/cdk-cli/src/sub_commands/receive.rs
@@ -1,14 +1,13 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::str::FromStr;
 use std::sync::Arc;
 
 use anyhow::{anyhow, Result};
-use cdk::amount::SplitTarget;
-use cdk::cdk_database::{Error, WalletDatabase};
-use cdk::nuts::{CurrencyUnit, SecretKey, Token};
+use cdk::cdk_database::{self, WalletDatabase};
+use cdk::nuts::SecretKey;
 use cdk::util::unix_time;
-use cdk::wallet::Wallet;
-use cdk::{Amount, UncheckedUrl};
+use cdk::wallet::multi_mint_wallet::MultiMintWallet;
+use cdk::Amount;
 use clap::Args;
 use nostr_sdk::nips::nip04;
 use nostr_sdk::{Filter, Keys, Kind, Timestamp};
@@ -35,9 +34,8 @@ pub struct ReceiveSubCommand {
 }
 
 pub async fn receive(
-    wallets: HashMap<UncheckedUrl, Wallet>,
-    seed: &[u8],
-    localstore: Arc<dyn WalletDatabase<Err = Error> + Sync + Send>,
+    multi_mint_wallet: &MultiMintWallet,
+    localstore: Arc<dyn WalletDatabase<Err = cdk_database::Error> + Send + Sync>,
     sub_command_args: &ReceiveSubCommand,
 ) -> Result<()> {
     let mut signing_keys = Vec::new();
@@ -62,10 +60,8 @@ pub async fn receive(
     let amount = match &sub_command_args.token {
         Some(token_str) => {
             receive_token(
+                multi_mint_wallet,
                 token_str,
-                wallets,
-                seed,
-                &localstore,
                 &signing_keys,
                 &sub_command_args.preimage,
             )
@@ -97,10 +93,8 @@ pub async fn receive(
             let mut total_amount = Amount::ZERO;
             for token_str in &tokens {
                 match receive_token(
+                    multi_mint_wallet,
                     token_str,
-                    wallets.clone(),
-                    seed,
-                    &localstore,
                     &signing_keys,
                     &sub_command_args.preimage,
                 )
@@ -128,29 +122,13 @@ pub async fn receive(
 }
 
 async fn receive_token(
-    token_str: &str,
-    wallets: HashMap<UncheckedUrl, Wallet>,
-    seed: &[u8],
-    localstore: &Arc<dyn WalletDatabase<Err = Error> + Sync + Send>,
+    multi_mint_wallet: &MultiMintWallet,
+    token: &str,
     signing_keys: &[SecretKey],
     preimage: &[String],
 ) -> Result<Amount> {
-    let token = Token::from_str(token_str)?;
-    let mint_url = token.proofs().iter().next().unwrap().0.clone();
-
-    let wallet = match wallets.get(&mint_url) {
-        Some(wallet) => wallet.clone(),
-        None => Wallet::new(
-            &mint_url.to_string(),
-            CurrencyUnit::Sat,
-            Arc::clone(localstore),
-            seed,
-            None,
-        ),
-    };
-
-    let amount = wallet
-        .receive(token_str, SplitTarget::default(), signing_keys, preimage)
+    let amount = multi_mint_wallet
+        .receive(token, signing_keys, preimage)
         .await?;
     Ok(amount)
 }