+++ crates/cdk-cli/src/sub_commands/restore.rs
@@ -1,8 +1,8 @@
-use std::collections::HashMap;
-
 use anyhow::{anyhow, Result};
+use cdk::nuts::CurrencyUnit;
 use cdk::url::UncheckedUrl;
-use cdk::wallet::Wallet;
+use cdk::wallet::multi_mint_wallet::WalletKey;
+use cdk::wallet::MultiMintWallet;
 use clap::Args;
 
 #[derive(Args)]
@@ -12,11 +12,15 @@ pub struct RestoreSubCommand {
 }
 
 pub async fn restore(
-    wallets: HashMap<UncheckedUrl, Wallet>,
+    multi_mint_wallet: &MultiMintWallet,
     sub_command_args: &RestoreSubCommand,
 ) -> Result<()> {
-    let wallet = wallets
-        .get(&sub_command_args.mint_url)
+    let wallet = multi_mint_wallet
+        .get_wallet(&WalletKey::new(
+            sub_command_args.mint_url.clone(),
+            CurrencyUnit::Sat,
+        ))
+        .await
         .ok_or(anyhow!("Unknown mint url"))?;
 
     let amount = wallet.restore().await?;